requirements.txt
yt-dlp>=2023.12.30
requests==2.31.0 
pydub>=0.25.1
python-dotenv>=1.0.0
google-generativeai>=0.3.0
